Add predicate to test if a type meets a requirement
Summary: I'm working on a system where every edge in the graph has a "requirement" of the input on the end of the edge imposed by the consumer; right now the only requirements are "exactly this type" and "any type less than or equal to this type". I've added a simple predicate to check whether a requirement is met; we'll use it in a subsequent diff.

Reviewed By: wtaha

Differential Revision: D22213381

fbshipit-source-id: 45c791c1ff2607b71f87196e08451e526dd577ea