Add inf types and edge requirements to graph debug output (#233)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/233

The type analyzer I am working on computes two things: first, what is the "infimum type" of every node in the accumulated graph? That is the answer to the question "if we had to convert this node to the smallest possible type in the lattice, what would that type be?"  That is the *greatest lower bound*, or *infimum*, of all types in the lattice, hence then name.

Second, we compute a *requirement* on each edge that is either "the value passed along this edge must be exactly this type" or "must be smaller than this type".

To help visualize this, I've added code to my graph builder to optionally display this information.

The way to read the annotation on a node, say, ">=N" is "this node can be converted to any type that is larger than or equal to Natural". The annotations on edges are "this edge must point **from** something of the stated type" -- that is, it is a requirement of the *pointed to* node.

{F242233215}

I made a few other minor changes along the way:

* Walid suggested a while back that using `float` to mean "real number" is slightly jarring to read. I've added `Real` as a synonym and will gradually modify the code to use it consistently.

* We're going to need in a few diffs a "human readable" string for each type and type requirement; since I was adding the abbreviated name for the graph builder now, I figured I'd add the long name too; it will be used in a few diffs down the road.

* Historically I've drawn graphs using the conventions of abstract syntax trees; the multiplication node has two arrows going out that point to its operands. But this is the opposite of the convention for Bayesian networks, where the inputs to the multiplication point at the multiplication. I've added an option to the graph builder to draw the arrows the "right" way, and will gradually change over the other test cases to use it.

Reviewed By: wtaha

Differential Revision: D22325570

fbshipit-source-id: 19634cf76e194448d61649824dfe63c472671f52