Makes RVIdentifier serializable using dill (#212)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/212

Introduces `__eq__` and `__hash__` on `RVIdentifier` which uses the id of `arguments` but the `__name__` and `__code__` of `function` (previously we were using id-based equality on both, and function id is not stable after deserialization).

Note that the same instance method on different instances will not be equal (as expected) because the first argument (`self`) will fail the id check for `arguments`.

Adds dependency/test case using dill to SerDe `MonteCarloSamples`. (dep is on pathos because it will be needed for replacing multiprocessing, which uses pickle internally)

See https://medium.com/emlynoregan/serialising-all-the-functions-in-python-cd880a63b591 for details on why we need dill.

Differential Revision: D21958354

fbshipit-source-id: d4357fc6590877095c4c0c7142c28378c32ce5fc