Start using "Real" when referring to BMG type (#237)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/237

We need objects to represent the types of the BMG type system (bool, natural, probability, positive real, real, tensor). I've been using "float" to represent "real" but it reads poorly. I've made an alias `Real` and started using that anywhere in the code where I am referring specifically to the BMG type, not the Python type.

Thanks to Walid for the suggestion.

Reviewed By: wtaha

Differential Revision: D22341483

fbshipit-source-id: bb00221bb39b707c773ff567f6c037c49c650be6