Start adding code for compiler error reporting
Summary:
This code is a start at tracking errors that we *expect to find in a program*, that must be reported to the user. It's not for errors where the compiler itself fails. Remember, error reporting is *not an exceptional case* for a compiler; programs contain errors.

Since I expect to have multiple kinds of errors I've made a base class and so far one derived class that will be for the scenario where a BMG graph cannot be generated because an edge requirement is always violated.  There will be more.

This code will be properly exercised in an upcoming diff.

Differential Revision: D22404528

fbshipit-source-id: bd632bb76c2681aff0e33fe4f17ac4fcb8f49a6e