Implement computation of "graph type" of each node.
Summary:
The `inf_type` of each node answers the question "what is the smallest BMG type system type I could convert this node to?"  I've added a new property `graph_type` which answers the question "what is the BMG type system type of this node right now?"

Since the BMG type system requires certain constraints to be met -- like, the type of a multiplication node is the type of its inputs, which must be the same -- there are some nodes where we cannot figure out the type. Those are given the "malformed" type.

The aim of the graph transformation which is coming up soon in this stack is: we wish to transform the graph so that there are no more malformed nodes, and all the requirements on the edges are met.

As a debugging aid I have added a visualizer for the graph type.

Differential Revision: D22399185

fbshipit-source-id: b938713ce2fb94887ebbe9a7b2addc8b314d9ac7