Enable new "fix problems" tool in end-to-end building of graph
Summary:
I've put my new "problem fixer" step in place before we attempt to generate a BMG graph. This replaces the previous prototype code for fixing type problems; it will be deleted in the next diff.

The new code does a better job than the old code; as you can see in the updated tests, it generates smaller graphs in some cases.

Right now if a problem is encountered, we throw an exception containing the error report. As we continue to evolve the compiler, this will be replaced with a more principled error reporting mechanism.

Differential Revision: D22416411

fbshipit-source-id: b6191593386a7053a4a4966102f86e3a1314c038