Create "problem fixer" module for transforming graphs to meet BMG requirements
Summary:
As we run the lifted program, the accumulated graph almost certainly does not meet the requirements of the BMG type system. This module attempts to fix the graph so that it does, by gradually transforming it until requirements are met; if requirements cannot be met, then a descriptive error message is added to an error report.

I have not yet added this new code to the code which generates the final graph; in a subsequent diff I will delete all the existing prototype graph transformation code and substitute in a call to fix_problems.

Differential Revision: D22405697

fbshipit-source-id: d76ebe9220281b7e42527f828724f106780d7cc9