implement simulate for batch predictives (#219)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/219

`simulate` generates forward samples from the model. This generates one observation per world (latent variable sample).  In the posterior predictive case, these samples are generated via vectorization.

Note this makes a change to `MonteCarloSamples` which accepts a list of dicts where each element corresponds to one chain as well as a single dict keyed on RVIdentifier and valued by a tensor whose leading dimension is the number of chains.

Reviewed By: nazanint

Differential Revision: D22042015

fbshipit-source-id: 3b2971e2ac90a0f08781d5059d660d77f7511f1b