fix bug in `single_inference_test` (#251)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/251

There was a bug in `single_inference_test`, which led to failures once in a while. This diff fixes the test-case.

The test-case was set to fail if `num_heads` was evaluated as 2 which it can be! (probably confused with the other summary `mean_value` which cannot have value of 2)

Reviewed By: wtaha

Differential Revision: D22491573

fbshipit-source-id: e9d0d07606f6e91a7962e823118ad56383a80023