fix bug in `single_inference_test` (#251)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/251

There was a bug in `single_inference_test`, which led to failures once in a while. This diff fixes the test-case.

The test-case was set to fail if `num_heads` was evaluated as 2 which it can be! (probably confused with the other summary `mean_value` which cannot have value of 2)

Reviewed By: wtaha

Differential Revision: D22491573

fbshipit-source-id: 8be45179f0ed3a5a19b8f2782a10488715a00ed1