Implement computation of "graph type" of each node.
Summary:
The `inf_type` of each node answers the question "what is the smallest BMG type system type I could convert this node to?"  I've added a new property `graph_type` which answers the question "what is the BMG type system type of this node right now?"

Since the BMG type system requires certain constraints to be met -- like, the type of a multiplication node is the type of its inputs, which must be the same -- there are some nodes where we cannot figure out the type. Those are given the "malformed" type.

The aim of the graph transformation which is coming up soon in this stack is: we wish to transform the graph so that there are no more malformed nodes, and all the requirements on the edges are met.

As a debugging aid I have added a visualizer for the graph type:

Each node is now annotated with `label : graph type >= inf type`. Consider the multiplication for instance. We have a multiplication of the tensor 0.5 (marked ":T >= P" because it *is* a tensor but it *could be* made into a probability or any larger type) by a sample from a beta, which is definitely a probability.  The multiplication is malformed because the input types are not the same, but we deduce that it *could* be made into a multiplication of probabilities or any larger type. So it is marked `: M >= P`

{F243030598}

Reviewed By: wtaha

Differential Revision: D22399185

fbshipit-source-id: bad10226736566ad35a5f4cdc91b59a9a9748e7c