Start adding code for compiler error reporting (#238)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/238

This code is a start at tracking errors that we *expect to find in a program*, that must be reported to the user. It's not for errors where the compiler itself fails. Remember, error reporting is *not an exceptional case* for a compiler; programs contain errors.

Since I expect to have multiple kinds of errors I've made a base class and so far one derived class that will be for the scenario where a BMG graph cannot be generated because an edge requirement is always violated.  There will be more.

This code will be properly exercised in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D22404528

fbshipit-source-id: eb2dd3ed72c93d415644e204b9787b36baf8909e