Rewrite rule to turning an empty keyword argument list into a single empty double starred argument (#250)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/250

This rewrite is part of a series of rewrites to implement the translation of function calls into single assignment normal form. This rule looks for calls like f(1,2,3) that lack any keyword arguments and turns them into f(1,2,3,**{}), which is a call with exactly one keyword argument, which is our planned normal form.

Note: This change requires updating a bunch of older tests, so, they were updated as well.

Reviewed By: ericlippert

Differential Revision: D22400524

fbshipit-source-id: a617168554b0e41143df4d603edce3762a4ffc89