Rewrite rule for turning a keyword argument into a double star argument (#249)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/249

This rewrite is part of a set of rewrites to implement the translation of function call arguments into single assignment form. This particular rewrite turns a simple keyword argument into a double star argument. Doing so places keyword arguments into the normal form for such arguments, namely, a single double star argument.

The change in translation of keyword arguments required some changes to the generation of BMG graph building code, which are made in bmg_to_bmg.py.

In implementing this rewrite it was noticed that there were some aspects of the implementation of the merging of two star arguments that needed improvement, and this was also done in this diff. For example, it is important that we do NOT merge arguments when the function itself is "dict", as this function also performs an additional dynamic check that is important to preserve. Also, merging arguments to "dict" by introducing an additional call to "dict" would cause an infinite loop.

Reviewed By: ericlippert

Differential Revision: D22295987

fbshipit-source-id: d28f77012a6bde4c8223ff76832b0a3f0d7da448