Checkpoint of current status of function call rewrites
Summary: This diff adds a simple testcase to confirm how function calls with no arguments get translated at this point. This test will need to be updated after the next rewrite rule is added, as it will mean that all function calls will have exactly two arguments when they are in normal form.

Differential Revision: D22396667

fbshipit-source-id: 36889046e2d721009695ddc848aa8564cdd8c7e4