Rewrite rule to turning an empty keyword argument list into a single empty double starred argument
Summary: This rewrite is part of a series of rewrites to implement the translation of function calls into single assignment normal form. This rule looks for calls like f(1,2,3) that lack any keyword arguments and turns them into f(1,2,3,**{}), which is a call with exactly one keyword argument, which is our planned normal form.

Differential Revision: D22400524

fbshipit-source-id: 386da71212ce7aa6c06863f488e3aab97d5d5f36