Checkpoint of current status of function call rewrites (#236)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/236

This diff adds a simple testcase to confirm how function calls with no arguments get translated at this point. This test will need to be updated after the next rewrite rule is added, as it will mean that all function calls will have exactly two arguments when they are in normal form.

Reviewed By: ericlippert

Differential Revision: D22396667

fbshipit-source-id: efd114ac5a224886d975771d6c0de9a6a087de0c