Rewrite for merging two double star arguments (#225)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/225

This is the second rewrite rule in a set of rules for handling keyword arguments in function calls.

Caveat: This diff is currently breaking a test in bm_to_bmg_test.py.  Because this regression is expected to be addressed to be handled by the next two rewrites to be added, this test is currently disabled.

Reviewed By: ericlippert

Differential Revision: D21650486

fbshipit-source-id: 98b7c66f0ad7a21118bda18accd0ebffdf7bea1b