Create "problem fixer" module for transforming graphs to meet BMG requirements (#240)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/240

As we run the lifted program, the accumulated graph almost certainly does not meet the requirements of the BMG type system. This module attempts to fix the graph so that it does, by gradually transforming it until requirements are met; if requirements cannot be met, then a descriptive error message is added to an error report.

I have not yet added this new code to the code which generates the final graph; in a subsequent diff I will delete all the existing prototype graph transformation code and substitute in a call to fix_problems.

Let's look at an example. Suppose we have accumulated this graph. (The annotations after the colon indicate the type of the node: T for tensor, P for probability, B for bool, M for malformed, R for real. Each edge has an annotation indicating the requirement that the pointed-to thing has on the pointed-from thing.)

{F243081559}

There are several problems here:

* The beta requires a positive real but it has a tensor for its input
* The multiplication is malformed because it is multiplying a tensor by a probability; it is required to have operands of the same type.
* The normal has two bad inputs; it requires a real and a positive real but it has a malformed multiplication and a tensor
* The Bernoulli requires a probability as its input but it has a malformed multiplication.

The fixer transforms it into this graph:

{F243081978}

* The now-unused tensor constants are orphaned and will not be emitted into the final graph.
* We've created new constants of the correct types for the inputs
* A ToReal node has been inserted on the edge between the normal and the multiplication.
* The multiplication now multiplies two probabilities

There is still a lot of work to do here in terms of detecting problems and giving good error messages, but we have now got a solid foundation on which to build that code.

Reviewed By: wtaha

Differential Revision: D22405697

fbshipit-source-id: 975067a89512c2ec1ec46b045efa0b082814283f