Generalized use of allowed_functions list (#242)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/242

A previous commit had introduced the variable allowed_functions to exclude certain function names from the back-end translation. The same exclusion seems appropriate in the single_assignment phase. For this reason, this list was moved to a new file and then also used to guide several rewrites in single_assignment.py.

Caveat: Currently Pyre is complaining about the import of the new file in single_assignment.py, but not bmg_graph_builder.py.

Reviewed By: ericlippert

Differential Revision: D22415511

fbshipit-source-id: 9c515cddebf4cc9eacbbdf541f4f3a9b4a59843f