Start implementing unsupported-node-type error reporting
Summary:
When we execute the model we accumulate a graph of the model which reflects the actual operations that occured in the Python program. However, not all of these operations have equivalents nodes in BMG.

Those that can be transformed into nodes that are compatible with BMG will be, by the "problem fixing" phase. However, some of them will not be fixable, and for those we will have to produce errors.

This diff does not produce those errors. Rather, it does two things:

* Adds a property to each node that indicates whether it is supported in BMG or not
* Throws an internal error if an attempt is made to generate a BMG graph from this node

Note that it is an *internal* error, which indicates that there is a bug; when we correctly implement the error detection pass in an upcoming diff, it should then be impossible for there to be any code path in which we attempt to generate a graph containing an unsupported node type.

Differential Revision: D22468742

fbshipit-source-id: 88e9625b1868616ef1654c852e583fc12a8e70d9