Add test case for fixable problem
Summary: We do not support division nodes in BMG yet, but we can fix a model that contains a division by a constant by turning it into a multiplication in a post-processing pass.  I haven't done that, but in the spirit of test-driven development I've written a test case that demonstrates that we currently produce an error message for this scenario. When we stop producing an error I'll update the test.

Differential Revision: D22490087

fbshipit-source-id: ee85bc6fb0a10f00a86a23e11f89c260b62ac9f4