Create test case for multiplication of bool by natural
Summary:
BMG does not permit multiplication of bool by natural to produce a natural; a multiplication must produce a probability, positive real, real or tensor.

However, we can construct models in which bools are multiplied by naturals and used as naturals, and there is a way to compile these models into a BMG that does meet the BMG rules: turn `bool * natural` into `if bool then natural else 0`, which is a legal natural in BMG.

I'm going to add this transformation to my problem fixer pass; when I do, this test case will stop throwing an exception.

Differential Revision: D22424087

fbshipit-source-id: d56bce5d22caf6ac5ed928be74a268971590a758