Implement graph transformation to eliminate division nodes
Summary:
BMG does not support a node representing division; if we have a model that contains a division by a constant, we rewrite it as a multiplication.

The test case illustrates how we start with a graph containing a division node, and then orphan that division node (and some constants) to produce a graph where the distributions depend only on the multiplication.

Differential Revision: D22490752

fbshipit-source-id: e8c19647857d778c1c5660f6e6ffdfa49d159fa5