Start implementing flat distribution
Summary:
BMG supports a "flat distribution" node, which is the standard uniform distribution (0.0 to 1.0).

I wish to distinguish the BMG node from the Uniform node in the Python model; the original graph accumulated should represent as closely as possible the Python program. We then transform that program, if possible, into graph nodes that meet the BMG rules.

In particular, suppose we have a Uniform(0.0, 1.0) in the original model. We could say, when we go to transform it into BMG form, that we just generate a flat distribution when we see this. But that would then leave the 0.0 and 1.0 nodes in the generated BMG graph *because those nodes are reachable from the uniform distribution node*.

The better choice is to replace the uniform distribution with a flat distribution, and orphan the 1.0 and 0.0 nodes.

This diff just introduces the type; I'll add support for building it in the graph builder in the next diff.

Differential Revision: D22539412

fbshipit-source-id: aaaf594effd4c715411b98a0cafe656dd17734f1