Create test case for multiplication of bool by natural (#245)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/245

BMG does not permit multiplication of bool by natural to produce a natural; a multiplication must produce a probability, positive real, real or tensor.

However, we can construct models in which bools are multiplied by naturals and used as naturals, and there is a way to compile these models into a BMG that does meet the BMG rules: turn `bool * natural` into `if bool then natural else 0`, which is a legal natural in BMG.

I'm going to add this transformation to my problem fixer pass; when I do, this test case will stop throwing an exception.

Reviewed By: wtaha

Differential Revision: D22424087

fbshipit-source-id: e16ded33e5fe55cc88c9d37efbe9f86bf18aeb06