Make notes on how to support different kinds of negation in graph translation
Summary:
We have a number of as-yet-unsolved problems in converting Python models to BMG graphs involving different kinds of negation; there is a bit of a mismatch between the Python semantics and the BMG semantics that we will need to bridge.

I've started here by making some notes in the code about our approach. I'll flesh these out over the next few diffs.

Reviewed By: wtaha

Differential Revision: D22447147

fbshipit-source-id: 5677fa07e941ad7c1010e85bf6062ddbaa70d93c