Add "unsupported operation" pass to problem fixing module
Summary:
We accumulate a graph representing the model as it executes; when it comes time to translate it into BMG, there can be nodes in the Python version of the model that have no representation in the BMG version.

In a future diff I will add code to remove as many of these as we can, and replace them with semantically equivalent nodes that we can represent. (For example, a division by a constant can be turned into a multiplication by a constant.)

However there will still be nodes that we cannot support until they are added to BMG. For those we produce an error message. This diff tests the production of those error messages.

I am not super thrilled with the wording, and it would be nice if in the future it referred back to the location in the source code, not the location in the graph. But it is a start!

Reviewed By: wtaha

Differential Revision: D22469906

fbshipit-source-id: 1e49e23b4a113c3eeb308903d8de5efa3fdeded5