Add test case for fixable problem
Summary: We do not support division nodes in BMG yet, but we can fix a model that contains a division by a constant by turning it into a multiplication in a post-processing pass.  I haven't done that, but in the spirit of test-driven development I've written a test case that demonstrates that we currently produce an error message for this scenario. When we stop producing an error I'll update the test.

Reviewed By: wtaha

Differential Revision: D22490087

fbshipit-source-id: ddabf3d4558dc6401de05da33a89161bfc886fd5