Add test case for fixable problem (#266)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/266

We do not support division nodes in BMG yet, but we can fix a model that contains a division by a constant by turning it into a multiplication in a post-processing pass.  I haven't done that, but in the spirit of test-driven development I've written a test case that demonstrates that we currently produce an error message for this scenario. When we stop producing an error I'll update the test.

Reviewed By: wtaha

Differential Revision: D22490087

fbshipit-source-id: 0f9ea4714e830bd6c3f7aaa2fe647e32cec28835