Add comments about power node support
Summary:
We will likely have to support a power node in BMG at some point; even if we do not, we'll need to come up with a strategy for translating power operators that appear in the original model into BMG nodes, which will require type analysis.

I've added some notes to the code on avenues we might pursue.

Reviewed By: wtaha

Differential Revision: D22515837

fbshipit-source-id: 10e39f47af186669e9f914cb43a2fef0a5db2575