Convert Uniform(0,1) distributions to flat BMG nodes (#275)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/275

A Python model can use a uniform distribution that is parameterized by arbitrary constants, tensors, or samples, but BMG only supports standard uniform distributions from 0.0 to 1.0 right now, which it calls a "flat" distribution on probabilities.

If we encounter a Uniform(0, 1) distribution, we now transform it into a flat distribution when compiling the model.

Reviewed By: wtaha

Differential Revision: D22553565

fbshipit-source-id: 109ff393e62c2dc1bf4ce8df68cd634efc12b2ac