Fix type system so EXP returns positive real
Summary:
In BMG and in the Beanstalk compiler, we treated EXP nodes as taking positive real, real or tensor, and returning whatever the input type was. But that's not quite right; an EXP node that takes a real can be known to return a positive real.

I've fixed that in both BMG proper and in the type analyzer in Beanstalk.

Differential Revision: D22650796

fbshipit-source-id: 025dcf87917ee6dcb6065d324371cb71df7a36e3