Finish adding support for chi2 distribution to BM->BMG compiler (#284)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/284

We now have end-to-end support for the chi2 distribution -- a special case of the gamma distribution -- in the BM->BMG compiler. If a model uses chi2, it is transformed into a gamma distribution when the final graph is created.

For example, suppose we start with this model:

        rv def hcs():
          return HalfCauchy(1.0)
        rv def chi2():
          return Chi2(hcs())

When we initially accumulate the graph we get:

{F244475089}

Which we cannot represent in BMG. But when we run the problem fixer, the constant real is replaced with a constant positive real, the chi2 is replaced with the equivalent gamma, and those original nodes become "orphans" that are not emitted into the final graph:

{F244475347}

Reviewed By: wtaha

Differential Revision: D22583490

fbshipit-source-id: f6a69d2992f6febeef2255b612306669e0f2c390