Fix uses of images in quick_start.md
Summary:
Docusaurus moves all images from docs/... into /img/ so you have to reference them from there - that's why the build was failing which is unrelated to static docs. So the fix for that is to use absolute references (/img/poisson.png) instead of relative (./poisson.png).

You could use the markdown image syntax for this, but I don't think it works with useBaseUrl - which is required for static docs to work, since it doesn't host the sites at the top level path.

So unfortunately you currently need to use img tags like this when using static docs.

Reviewed By: michaeltingley

Differential Revision: D22863621

fbshipit-source-id: 8066f16d5c824c220f66ad95b80c1a39db6200f2