Start introducing new class for representing lattice types
Summary:
In the original version of the graph type analyzer, I just used actual Python types -- float, bool, and so on -- to represent different types in the graph.

This is first, a somewhat questionable decision to use those objects as markers for what is essentially a different type system, and second, since BMG is now supporting more complex types such as matrix types, this will no longer be sufficient. We need proper classes that represent the BMG type lattice in a principled way.

However we have a LOT of code already in the product that assumes that "type" is the type of BMG lattice types. To start to fix that, I'm first going to simply make an alias *for the purposes of type annotations in the source code only*. Once we've got all the source code annotations correct, we can then refactor into the new type system without too much disruption.

Differential Revision: D23015466

fbshipit-source-id: 060b739d40676990c60d47cb52e64b7c00a2f6d8