Start replacing uses of bool type with proper marker type
Summary:
As noted in previous diffs in this stack, the current Beanstalk type analyzer uses real Python types like float and bool as markers for the types of graph nodes. We are going to need to abandon this practice as we start to support matrix types. The first step to make this happen is I am replacing usages of "bool" with an alias Boolean, that will soon be a proper object representing a particular lattice type.

So far I am just updating the type definition and its tests. I'll make this update throughout the rest of the system in later diffs, along with also adding the BMGLatticeType annotation mentioned earlier.

Differential Revision: D23015620

fbshipit-source-id: 23120040d846160bb12fc5757d60f6c631061058