Continue renaming refactorings in beanstalk compiler
Summary: As noted in previous diffs in this stack, we are revamping the BMG type system to support matrix types, and that means revamping the type analysis in Beanstalk. We need to stop using "type" as the type of lattice types, and instead create specific objects representing nodes in the lattice, and we need to stop using "bool" as a marker for Boolean nodes.  Right now I've just made aliases; once all the aliases and type annotations are correct, we can start turning on new features in the type system and test them out.

Differential Revision: D23016080

fbshipit-source-id: 4685094c1ba28c4ab3524b581b158686458eea6e