Rewrite for list comprehensions into SSA form (#337)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/337

List comprehensions are expressions like

   y = [c for v_i in e_i if b_i].

The key insight in this diff is that such expressions can be desugared
into expressions that use operators for which we already have
rewrites. In particular, an expression such as the above would be
translated into one with the following form:

   def p():
      r = []
      for v_i in e_i
         if b_i:r.append(c)
      return r
   y=p()

From then on, existing rewrites handle the rest of SSA conversion.

We use a new function definition in the rewriten term to address
scoping issues, of which there are two. First, we do not want v_i to
be visible outside the scope of the lhs of the assignment we started
off with (actually, not even outside the comprehension brackets within
that lhs). Second, we do not want any interference between the names
v_i and the name y, which can in principle overlap. Both issues are
addressed with the use of the new function definition. Note also that
p and r are fresh names.

While adding this rewrite, a couple of incidental additions were
made. First, some comments were added to existing functions. Second, a
new function freshName was introduced to make it more convenient to
introduced fresh names (such as p and r) while building a new term.

Caveat: We do not currently deal with the possible SYNC annotation on generators.

Reviewed By: ericlippert

Differential Revision: D23247448

fbshipit-source-id: 48131d64cf37ad732d4a97dd6fc516767e6e22af