Make a type for lattice types
Summary: The BMG type system now has complex types such as 2x2-matrix-of-reals, and so on; we will need a proper class for representing instances of BMG types. In this diff I create a base type, but we are still only handling "scalar" types. I will gradually add support for matrix types in subsequent diffs

Reviewed By: wtaha

Differential Revision: D23116114

fbshipit-source-id: 6c5fc3b11a69c6c6c13a6d868733765c092b2bea