Make a type for lattice types
Summary: The BMG type system now has complex types such as 2x2-matrix-of-reals, and so on; we will need a proper class for representing instances of BMG types. In this diff I create a base type, but we are still only handling "scalar" types. I will gradually add support for matrix types in subsequent diffs

Differential Revision: D23116114

fbshipit-source-id: 942c14290f4e92fb7be71b9897d7da257f0c7fd6