Make a type for lattice types (#340)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/340

The BMG type system now has complex types such as 2x2-matrix-of-reals, and so on; we will need a proper class for representing instances of BMG types. In this diff I create a base type, but we are still only handling "scalar" types. I will gradually add support for matrix types in subsequent diffs

Reviewed By: wtaha

Differential Revision: D23116114

fbshipit-source-id: 9ec9bc60d3dc628368327006f87c83bc997b02ec