Refactorings to simplify fresh names function and minor corrections (#349)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/349

The goal of this refactoring is to make the fresh names
function easier to use. The main change is to make it take a list of
names, and to be explicit at the use location about the version of the
new name that we would like to use, namely, either "load" or "store".

Minor fixes includ changing freshNames to fresh_names to be consistent
with style in rest of code, and to introduce some local variables in
the listComp rewrite to help reduce the maximum nesting depth of the
code.

Hopefuly, the resulting code is easier to read than the original.

Reviewed By: ericlippert

Differential Revision: D23351942

fbshipit-source-id: 0bb25d65bbbf193612389ec86d5fab7ec838693f