Let user_install run all tests on CircleCI (#348)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/348

Previously, user install was tested using `pytest --pyargs beanmachine`, which only run tests that are part of the `beanmachine` package. As a result, integration tests that are under a non-package `tests/` directory are not being run.

By making the change, we are using pytest to discover all test files that are under the project's root directory ([see how pytest matches pattern here](https://docs.pytest.org/en/stable/reference.html#confval-python_files)) and run them as independent files. Any `import beanmachine` in those tests files will import the version that user has already installed, which ensure that we are still testing the correct package. In addition, this means that the tests do not necessarily have to be installed as part of the package.

Reviewed By: nimar

Differential Revision: D23350008

fbshipit-source-id: 9945e2130660e4a838b0f1c800f15523962fcf5e