Start tracking "one-hot" scalars in type system (#353)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/353

The addition of both Boolean and row-simplex matrices to the type system presents a problem.

A "one-hot" matrix has every row all 0 except for a single 1. But that is clearly usable as both a Boolean matrix -- since every element is 0 or 1 -- or as a row-simplex matrix -- since every row's values are between 0 and 1, and they add to 1.

The solution is to add "one-hot matrix" as a lattice type smaller than both Boolean matrix and simplex matrix.  This lattice type only shows up in type analysis, and only in cases where we have a constant value; it is not an actual type in the BMG type system proper.

Since we are treating scalars the same as 1x1 matrices, we can start testing support of one-hot by treating any constant 1, 1.0, or True, or any tensor containing a single element of those values, as having the infimum type "One" which I will make a synonym for "1x1 one-hot matrix".

As we add support in the analysis for larger matrix types we will add tests for larger one-hot matrices.

Reviewed By: wtaha

Differential Revision: D23193024

fbshipit-source-id: f10573a1bab67cc1438a07bdc07cc9aed8c039ed