Fix type system so EXP returns positive real
Summary:
In BMG and in the Beanstalk compiler, we treated EXP nodes as taking positive real, real or tensor, and returning whatever the input type was. But that's not quite right; an EXP node that takes a real can be known to return a positive real.

BMG has been fixed so that EXP only returns positive real. I've updated Beanstalk to match.

Reviewed By: wtaha

Differential Revision: D22650796

fbshipit-source-id: b973c78c93a01b7c0ad5444514208b3a0959bce3