Fix type system so EXP returns positive real (#357)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/357

In BMG and in the Beanstalk compiler, we treated EXP nodes as taking positive real, real or tensor, and returning whatever the input type was. But that's not quite right; an EXP node that takes a real can be known to return a positive real.

BMG has been fixed so that EXP only returns positive real. I've updated Beanstalk to match.

Reviewed By: wtaha

Differential Revision: D22650796

fbshipit-source-id: cf2b39a99f9de524d9e2ef28247668dd12c2bc09