Add NEGATIVE_LOG operator to BMG (#359)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/359

Nim suggested to me a while back that we add a NEGATIVE_LOG operator to BMG; the benefit of such an operator is that we know that if given a probability, it will produce a positive real. This then means we get good type analysis without having to introduce the concept of "negative real".

Reviewed By: nimar

Differential Revision: D23437112

fbshipit-source-id: 4bfea6ef86168aec7a4d64c849d976812bf1b03d