Start adding support for BMG complement node (#364)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/364

BMG supports two kinds of negation: unary minus, which operates on reals, and complement, which has the semantics of "not b" for a Boolean operand and "1-p" for a probability operand. (Though the keen-eyed reviewer will notice that "1-b" and "not b" are the same thing in Python if b is a bool.)

My plan is to transform a Python model's uses of "not b" and "1-p" to the complement node, so we'll add such a node to the node hierarchy in the compiler. We do not make use of it yet; I will add the transformations in a later diff.

Reviewed By: wtaha

Differential Revision: D23328794

fbshipit-source-id: 48e93776878b8f4e46cc38ea1b4e3f394ce2f4d4