Mark conjugate tests as nightly and schedule CircleCI workflow (#362)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/362

Currently, many of our unit tests are not "unit" and take relatively long time to run. Because it would be time-consuming to run such expensive tests for every change we make, it would be better if we mark some of those "unit tests" as nightly tests that run once per day.

Conjugate tests are one of those cases that require high timeout to run. In this diff, all of the test files for conjugate tests are renamed to add the `*_nightly.py` suffix, which is used to denote test files that should be run on a daily basis and will be excluded from unit test runs.

In addition, the config file for CircleCI is also updated, which will schedule a new `nightly_test_py37_pip` on 00:00 UTC (12:00 am UTC) each day and will run all tests with `*_nightly.py` suffix as an ordinary user (i.e., in user install mode) on `master` branch only.

Reviewed By: nimar

Differential Revision: D23459213

fbshipit-source-id: 14a7e0245c126d6d2e15fb2ec29b152fcb683b65