Create test showing (1-p) is not (yet) a probability (#365)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/365

In BMG, for a probability-valued node p, today we represent 1-p as 1+(-p).  But the negate operator is real valued, and therefore 1-p is also real valued, not probability valued. The newly added test demonstrates that fact.

In a subsequent diff I'll be fixing this problem, so that 1-p is instead generated as complement(p), which is probability valued; I will update the test then.

Reviewed By: wtaha

Differential Revision: D23386801

fbshipit-source-id: db749a8cd6cc7e115d5469ddd37d37bad216d531