Convert 1+(-p) to complement node (#367)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/367

In BMG there is no subtraction node but there is a "complement" node which takes as its input a probability between 0.0 and 1.0, and produces 1-p as its output.

Negation is defined as taking a real, so 1 + (-p) by default will be of type real if we transform it directly into BMG. But if we detect this scenario and turn it into a complement, we can keep the type as probability, a smaller type than real.

Here is the final graph showing both included and orphaned nodes after the problem fixing phase. As you can see, we orphan the 2.0 "real" constant in favor of a 2.0 "positive real" constant, and the original "1 + (-beta)" node is orphaned in favor of a complement(beta) node; since the 1+(-beta) node no longer flows into a sample or observation it will be omitted when we emit the BMG.  I'll show that in a test case in a later diff.

{F312524119}

Reviewed By: wtaha

Differential Revision: D23403717

fbshipit-source-id: 4d97639fb86d07d230277b31c3f1a6eda4daaa67