Minor updates to problem fixer
Summary:
Since we have eliminated the Tensor type in BMG, we'll have some work to do later:

(1) Give helpful error messages when we have deduced that the minimal type of a node is Tensor, which we now use to mean ">2-dimensional matrix".

(2) Correctly handle cases where the type of a node is a 1- or 2-dimensional matrix.

I've updated comments describing this work, and also fixed a small bug caused by the introduction of the One type.

Reviewed By: wtaha

Differential Revision: D23387094

fbshipit-source-id: afcc5397c84fb7c359bfb18b0e4031a7ff5aaeb5