Minor updates to problem fixer (#368)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/368

Since we have eliminated the Tensor type in BMG, we'll have some work to do later:

(1) Give helpful error messages when we have deduced that the minimal type of a node is Tensor, which we now use to mean ">2-dimensional matrix".

(2) Correctly handle cases where the type of a node is a 1- or 2-dimensional matrix.

I've updated comments describing this work, and also fixed a small bug caused by the introduction of the One type.

Reviewed By: wtaha

Differential Revision: D23387094

fbshipit-source-id: 7de6238c23e04ba54d634526bbde1a045f36bb77