Start adding support for negative log to compiler
Summary:
BMG now supports a negative-log node; by generating such a node for -log(x), we can deduce that the value is a positive real if the operand is a probability.

So far I've just implemented the node and verified that its participation in the type system is as expected. In the next few diffs I will add the transformation.

Reviewed By: wtaha

Differential Revision: D23463817

fbshipit-source-id: 8688fb77aafb12cb57d9dd3a2c1c84f8440fdd79