Start adding support for negative log to compiler (#370)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/370

BMG now supports a negative-log node; by generating such a node for -log(x), we can deduce that the value is a positive real if the operand is a probability.

So far I've just implemented the node and verified that its participation in the type system is as expected. In the next few diffs I will add the transformation.

Reviewed By: wtaha

Differential Revision: D23463817

fbshipit-source-id: efe46917537c981e0da92b62d95f930364c9cfb5