Convert -log(x) nodes to negative_log(x) (#371)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/371

If we detect usages of -log(x) and compile them into the new "negative log" node, then we can get a tighter bound on types; specifically, if we know that x is a probability then we know that negative_log(x) is a positive real.

This diff implements the fixer that does the graph transformation and updates the test.

I'll add an end-to-end test demonstrating this in the next diff.

Reviewed By: wtaha

Differential Revision: D23485199

fbshipit-source-id: 6c481e9aa6383d10c554e4b95c94d56e71f53cdb