add dependency on pybind11 and remove pytorch dep (#372)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/372

bmgraph still had a dependency on pytorch through the CppExtension module. Remove this dependency but as a result needed to add pybind11 dependency which was added here.

Also removed the flags to specify header files this was getting too long for each dependency. On Linux these dependencies are already in /usr/include so we don't need anything special.

In Conda and on Windows the user can specify the include_dirs flag to build_ext. For example:

    pip install . --global-option=build_ext --global-option="-Isrc:<eigen-location>/include/eigen3:<pybind11-location>/include:<libboost-location>/include"

This is also documented in this quip doc https://fb.quip.com/5UbDA2aOmjkU

Reviewed By: horizon-blue

Differential Revision: D23638362

fbshipit-source-id: 61d3ce71aea1854d3a1ff74704dbc4544a0ed75e