Add test case for compilation of linear regression model (#377)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/377

We're at a point now where it makes sense to start making test cases for specific real-world models that we'd like to compile from BeanMachine to BMGraph. The first one: linear regression.

We did not previously have a debugging helper for "show this graph after it is transformed to BMG as DOT", so I've added that helper. The compiled, type-annotated graph of this model looks like this:

{F329459015}

Reviewed By: wtaha

Differential Revision: D23590343

fbshipit-source-id: 8ef88f2f80a925ad39b8c45d1ab2bb56a4ae4ff9