End to end test of unvectorized logistic regression
Summary:
As noted in the previous diff, I'd like to start making end to end tests that show we can compile realistic models to BMG. Here we have an unvectorized version of the Bayesian logistic regression model.

The graph of the unvectorized model looks a bit of a mess but if you trace it through you'll see it is actually quite straightforward.

{F333485030}

Reviewed By: wtaha

Differential Revision: D23614354

fbshipit-source-id: e53c524f0618ceaa556e5371136fe13984aea43c