Fix bad observations
Summary: The observed value of a sample must be a value that the distribution can actually produce. Here we add a pass to the problem fixer which looks for observations that do not match the distribution type; those that can be fixed are fixed, and those that cannot produce an error.

Reviewed By: wtaha

Differential Revision: D23697546

fbshipit-source-id: 049065862f516b8edc3546230e0c609af45e9b92