Fix bad observations (#381)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/381

The observed value of a sample must be a value that the distribution can actually produce. Here we add a pass to the problem fixer which looks for observations that do not match the distribution type; those that can be fixed are fixed, and those that cannot produce an error.

Reviewed By: wtaha

Differential Revision: D23697546

fbshipit-source-id: d67d080b40ecac10036dd24273ae5f83c4c248bc