Infer posteriors from Python models using BMG
Summary:
At long last, we can actually do a true end-to-end test where we have a model expressed in Python source code, and we get inferred samples from BMG.

The mechanisms I've created to do this have some pretty rough edges, but this proof-of-concept shows that we can make it work; we can decide later how best to expose this ability to actual users.

Right now it works like this:

* Write your model in Python in a single module
* At the bottom of your model, create two global variables "queries" and "observations" -- just as you would in a regular Python model; the difference is that they must be in variables with those names.
* Pass the whole model as a string to my new "infer" method.
* The infer method passes back a list of samples for the queried stochastic quantities.

The implementation details of the infer method are:

* The string is parsed as a Python program and transformed such that when it executes, it accumulates nodes into a graph builder
* Once execution has finished, we add observations and queries to the graph
* The graph is then transformed into a legal BMG graph which is created via Python-to-native bindings
* We call infer() on the BMG graph and return the result.

The test case is the standard "hello world" model where we create a coin with a prior on its fairness, observe some coin flips, and then get a posterior on the fairness. In this case the prior fairness is Beta(2,2), so it is most likely to be 50% fair. We observe three tails and one head, which is evidence that the coin is really 25% fair. The posterior should then be in the middle; the average of the samples from the posterior is around 37% fair.

Reviewed By: wtaha

Differential Revision: D23695270

fbshipit-source-id: 52398b779403db22a90f31ceff6ef24b1a373293